requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
docopt==0.6.2
filelock==3.0.12
idna==2.10
isort==5.7.0
pip-api==0.0.19
pipreqs==0.4.10
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
virtualenv==20.4.2
yarg==0.1.9
